+++ pymc3/variational/callbacks.py
@@ -1,8 +1,11 @@
+import collections
+
 import numpy as np
 
 __all__ = [
     'Callback',
-    'CheckParametersConvergence'
+    'CheckParametersConvergence',
+    'Tracker'
 ]
 
 
@@ -76,3 +79,53 @@ class CheckParametersConvergence(Callback):
     @staticmethod
     def flatten_shared(shared_list):
         return np.concatenate([sh.get_value().flatten() for sh in shared_list])
+
+
+class Tracker(Callback):
+    """
+    Helper class to record arbitrary stats during VI
+
+    It is possible to pass a function that takes no arguments
+    If call fails then (approx, hist, i) are passed
+
+
+    Parameters
+    ----------
+    kwargs : key word arguments
+        keys mapping statname to callable that records the stat
+
+    Examples
+    --------
+    Consider we want time on each iteration    
+    >>> import time
+    >>> tracker = Tracker(time=time.time)
+    >>> with model:
+    ...     approx = pm.fit(callbacks=[tracker])
+    Time can be accessed via :code:`tracker['time']` now
+
+    For more complex summary one can use callable that takes
+    (approx, hist, i) as arguments
+    >>> with model:
+    ...     my_callable = lambda ap, h, i: h[-1]
+    ...     tracker = Tracker(some_stat=my_callable)
+    ...     approx = pm.fit(callbacks=[tracker])
+    """
+    def __init__(self, **kwargs):
+        self.whatchdict = kwargs
+        self.hist = collections.defaultdict(list)
+
+    def record(self, approx, hist, i):
+        for key, fn in self.whatchdict.items():
+            try:
+                res = fn()
+            except TypeError:
+                res = fn(approx, hist, i)
+            self.hist[key].append(res)
+
+    def clear(self):
+        self.hist = collections.defaultdict(list)
+
+    def __getitem__(self, item):
+        return self.hist[item]
+
+    __call__ = record
