+++ pymc3/variational/inference.py
@@ -2,17 +2,15 @@ from __future__ import division
 
 import logging
 import warnings
-import collections
-import tqdm
 
 import numpy as np
+import tqdm
 
 import pymc3 as pm
+from pymc3.variational import test_functions
 from pymc3.variational.approximations import MeanField, FullRank, Empirical
 from pymc3.variational.operators import KL, KSD, AKSD
 from pymc3.variational.opvi import Approximation
-from pymc3.variational import test_functions
-
 
 logger = logging.getLogger(__name__)
 
@@ -22,7 +20,6 @@ __all__ = [
     'SVGD',
     'ASVGD',
     'Inference',
-    'Summary',
     'fit'
 ]
 
@@ -193,56 +190,6 @@ class Inference(object):
         finally:
             progress.close()
         self.hist = np.concatenate([self.hist, scores])
-
-
-class Summary(object):
-    """
-    Helper class to record arbitrary stats during VI
-
-    It is possible to pass a function that takes no arguments
-    If call fails then (approx, hist, i) are passed
-
-
-    Parameters
-    ----------
-    kwargs : key word arguments
-        keys mapping statname to callable that records the stat
-
-    Examples
-    --------
-    Consider we want time on each iteration    
-    >>> import time
-    >>> summary = Summary(time=time.time)
-    >>> with model:
-    ...     approx = pm.fit(callbacks=[summary])
-    Time can be accessed via :code:`summary['time']` now
-
-    For more complex summary one can use callable that takes
-    (approx, hist, i) as arguments
-    >>> with model:
-    ...     my_callable = lambda ap, h, i: h[-1]
-    ...     summary = Summary(some_stat=my_callable)
-    ...     approx = pm.fit(callbacks=[summary])
-    """
-    def __init__(self, **kwargs):
-        self.whatchdict = kwargs
-        self.hist = collections.defaultdict(list)
-
-    def record(self, approx, hist, i):
-        for key, fn in self.whatchdict.items():
-            try:
-                res = fn()
-            except TypeError:
-                res = fn(approx, hist, i)
-            self.hist[key].append(res)
-
-    def clear(self):
-        self.hist = collections.defaultdict(list)
-
-    def __getitem__(self, item):
-        return self.hist[item]
-
-    __call__ = record
 
 
 class ADVI(Inference):
